Move class `HookHelper`
