+++ app/src/main/java/andhook/ui/MainApplication.java
@@ -6,6 +6,7 @@ import android.util.Log;
 public final class MainApplication extends Application {
     @Override
     public void onCreate() {
+        super.onCreate();
         Log.i(MainApplication.class.toString(), "onCreate");
         andhook.test.AndTest.RunTest(this, this.getContentResolver());
     }
