+++ app/src/main/jni/Application.mk
@@ -1,4 +1,4 @@
-APP_ABI := armeabi-v7a x86
+APP_ABI := armeabi-v7a arm64-v8a x86 x86_64
 APP_PLATFORM := android-27
 APP_STL := stlport_static
 APP_OPTIM := release
